refactored to split components to individual files
