+++ app/input-label-view.tsx
@@ -1,0 +1,5 @@
+export const InputLabelView: React.FC<{ text: string }> = ({
+    text
+}) => {
+    return <p className="pb-2"> {text} </p>
+};
