+++ app/utility.ts
@@ -1,8 +1,8 @@
-import { FormOption } from "./interface"
+// import { FormOption } from "./interface"
 
-export const getKeyName = (item:FormOption) => {
-    return item.category + '__' + item.name
-}
+// export const getKeyName = (item:FormOption) => {
+//     return item.category + '__' + item.name
+// }
 
 export const validateInput = (data: any, schema: any, setErrorFunc: any) => {
     schema.validate(data, { abortEarly: false }).then(
